EZP-23897: Moved event subscribers to eZ\Publish\Core\MVC\Symfony\Cache\Http\EventListener
